+++ platform/platform.json
@@ -37,6 +37,25 @@
             }
         },
         {
+            "apiVersion": "apiextensions.k8s.io/v1beta1",
+            "kind": "CustomResourceDefinition",
+            "metadata": {
+                "name": "schedulerrecords.elastic.io"
+            },
+            "spec": {
+                "group": "elastic.io",
+                "names": {
+                    "kind": "SchedulerRecord",
+                    "listKind": "SchedulerRecordList",
+                    "plural": "schedulerrecords",
+                    "shortNames": [],
+                    "singular": "schedulerrecord"
+                },
+                "scope": "Namespaced",
+                "version": "v1"
+            }
+        },
+        {
             "apiVersion": "v1",
             "kind": "ServiceAccount",
             "metadata": {
@@ -106,7 +125,8 @@
                         "elastic.io"
                     ],
                     "resources": [
-                        "flows"
+                        "flows",
+                        "schedulerrecords"
                     ],
                     "verbs": [
                         "create",
@@ -226,13 +246,17 @@
                                         "value": "amqp://guest:guest@rabbitmq-service.platform.svc.cluster.local"
                                     },
                                     {
+                                        "name": "RABBITMQ_MANAGEMENT_URI",
+                                        "value": "http://guest:guest@rabbitmq-service.platform.svc.cluster.local:15672/"
+                                    },
+                                    {
                                         "name": "SELF_URL",
                                         "value": "http://api-service.platform.svc.cluster.local"
                                     }
                                 ],
                                 "terminationMessagePath": "/dev/termination-log",
                                 "terminationMessagePolicy": "File",
-                                "imagePullPolicy": "Always",
+                                "imagePullPolicy": "IfNotPresent",
                                 "securityContext": {
                                     "privileged": false
                                 },
@@ -331,7 +355,7 @@
                                 ],
                                 "terminationMessagePath": "/dev/termination-log",
                                 "terminationMessagePolicy": "File",
-                                "imagePullPolicy": "Always",
+                                "imagePullPolicy": "IfNotPresent",
                                 "livenessProbe": {
                                     "httpGet": {
                                         "path": "/healthcheck",
@@ -408,7 +432,7 @@
                                 ],
                                 "terminationMessagePath": "/dev/termination-log",
                                 "terminationMessagePolicy": "File",
-                                "imagePullPolicy": "Always",
+                                "imagePullPolicy": "IfNotPresent",
                                 "livenessProbe": {
                                     "httpGet": {
                                         "path": "/healthcheck",
