+++ services/communication-router/config/default.json
@@ -1,0 +1,1 @@
+{}
