+++ services/flows-operator/config/.gitignore
@@ -1,0 +1,1 @@
+dev.json
