+++ services/flows-operator/config/default.json
@@ -1,0 +1,3 @@
+{
+    "SELF_API_URI": "http://api-service.platform.svc.cluster.local:1234"
+}
