+++ services/flows-operator/package.json
@@ -4,16 +4,16 @@
   "description": "Standalone version of elastic.io platform",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\"",
+    "lint": "eslint ./index.js ./src",
+    "test": "npm run lint",
     "start": "node index.js"
   },
   "author": "elastic.io",
   "license": "APL-2.0",
   "dependencies": {
-    "JSONStream": "1.3.1",
     "express": "4.16.3",
     "lib": "file:../lib",
-    "lodash": "^4.17.11",
+    "lodash": "4.17.11",
     "node-uuid": "1.4.7",
     "rabbitmq-stats": "1.1.2"
   },
