+++ services/lib/data/SchedulerCRD.json
@@ -1,0 +1,19 @@
+{
+    "apiVersion": "apiextensions.k8s.io/v1beta1",
+    "kind": "CustomResourceDefinition",
+    "metadata": {
+        "name": "schedulerrecords.elastic.io"
+    },
+    "spec": {
+        "group": "elastic.io",
+        "names": {
+            "kind": "SchedulerRecord",
+            "listKind": "SchedulerRecordList",
+            "plural": "schedulerrecords",
+            "shortNames": [],
+            "singular": "schedulerrecord"
+        },
+        "scope": "Namespaced",
+        "version": "v1"
+    }
+}
