+++ services/lib/index.js
@@ -1,16 +1,9 @@
-const QueueCreator = require('./QueueCreator.js');
-const CRD = require('./CRD.json');
-const Flow = require('./Flow.js');
-const App = require('./App.js');
-const Config = require('./Config.js');
-const AMQPService = require('./AMQPService.js');
-const K8sService = require('./K8sService.js');
 module.exports = {
-    QueueCreator,
-    CRD,
-    Flow,
-    App,
-    AMQPService,
-    K8sService,
-    Config
+    App: require('./src/App.js'),
+    Flow : require('./src/Flow.js'),
+    QueueCreator: require('./src/QueueCreator.js'),
+    AMQPService: require('./src/AMQPService.js'),
+    K8sService: require('./src/K8sService.js'),
+    FlowCRD: require('./data/FlowCRD.json'),
+    SchedulerCRD: require('./data/SchedulerCRD.json')
 };
