+++ services/scheduler/Dockerfile
@@ -2,6 +2,8 @@ FROM node:8-alpine AS base
 WORKDIR /usr/src/scheduler
 COPY lib ../lib
 COPY scheduler/package.json scheduler/package-lock.json scheduler/index.js ./
+COPY scheduler/src ./src
+COPY scheduler/config/default.json ./config/default.json
 
 # Image for building and installing dependencies
 # node-gyp is required as dependency by some npm package
@@ -10,9 +12,11 @@ COPY scheduler/package.json scheduler/package-lock.json scheduler/index.js ./
 FROM base AS dependencies
 RUN apk add --no-cache make gcc g++ python
 RUN npm install --production --quiet
+RUN cd ../lib && npm install --production --quiet
 
 FROM base AS release
 COPY --from=dependencies /usr/src/scheduler/node_modules ./node_modules
+COPY --from=dependencies /usr/src/lib/node_modules ../lib/node_modules
 RUN rm -rf package-lock.json
 CMD ["npm", "start"]
